pyproject.toml
[project]
name = "indolj-sales-narration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.1",
    "fastapi>=0.122.0",
    "fastapi-cache>=0.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-deepseek>=1.0.1",
    "langchain-experimental>=0.4.0",
    "langchain-google-genai>=4.0.0",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "langchain[google-genai]>=1.1.0",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.4",
    "prophet>=1.2.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.51.0",
    "uvicorn[standard]>=0.38.0",
]
